Force geocoder_coalesce_radius to be a number
In some cases when reading metadata from mbtiles sources parameters that
should be numeric, such as geocoder_coalesce_radius are converted to
strings. This leads to errors like this when instantiating a GridStore:

    TypeError: invalid type: string "400", expected f64

The root problem of parameters changing types still needs to be tracked
down, but this approach mitigates the problem for now.